Maintain power-of-two vector widths
There was a bug in the `post-reset` function that only surfaced with our
Rosette 4 changes. `post-reset` replaces all verified FIFO memories with
a new zero vector with length equal to the depth of the FIFO (or depth-1
in the case of async FIFOs). The problem with this is that FIFOs are
indexed with bitvectors, and if the length of this vector isn't a power
of two, the `vector-ref-bv` call that implements `select` in the updated
version of rtl/yosys will emit side assertions when given a symbolic
index (to check that the index isn't too large).

Since Yosys gets around this problem by rounding up all memory sizes to
a power of 2 anyways, the solution is to maintain these correct sizes
(and to make this easy to get right, I switched `post-reset` to check the
length of the vector instead of hard-coding it). I also had to modify
the async FIFO proof to handle an edge case related to this issue.